backend/requirements.txt
aniso8601==9.0.1
attrs==20.3.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
flask-restplus==0.13.0
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
mongoengine==0.23.0
pycparser==2.20
PyJWT==2.0.1
pymongo==3.11.3
pyrsistent==0.17.3
pytz==2021.1
six==1.15.0
Werkzeug==0.16.1
WTForms==2.3.3
